Remove cyclic imports
